requirements.txt
requests
python-dotenv
intervalsicu
pandas
matplotlib
seaborn